+++ amworkflow/src/core/workflow.py
@@ -1,353 +1,392 @@
-from OCC.Core.TopoDS import TopoDS_Shape
-import sys
-import os
-from amworkflow.src.constants.enums import Directory as D
-from amworkflow.src.constants.enums import Label as L
-from amworkflow.src.utils.parser import yaml_parser, cmd_parser
-import gmsh
-import amworkflow.src.infrastructure.database.engine.config as CG
-from amworkflow.src.infrastructure.database.models.model import XdmfFile, H5File, FEResult, SliceFile, GCode, ModelProfile, ModelParameter, GeometryFile
-from amworkflow.src.infrastructure.database.cruds.crud import insert_data, query_multi_data, delete_data
-from amworkflow.src.geometries.mesher import mesher
-from amworkflow.src.utils.writer import mesh_writer, mk_dir
-from amworkflow.src.utils.permutator import simple_permutator
-from amworkflow.src.utils.writer import namer, stl_writer, batch_num_creator
-import numpy as np
-from amworkflow.src.utils.db_io import downloader
-from amworkflow.src.constants.data_model import MapParamModel, DeepMapParamModel
-from amworkflow.src.utils.sanity_check import path_valid_check, dimension_check
-from amworkflow.src.constants.exceptions import NoDataInDatabaseException, InsufficientDataException
-from amworkflow.src.utils.reader import get_filename
-from amworkflow.src.utils.reader import step_reader, stl_reader
+#self dependencies
 from amworkflow.src.interface.cli.cli_workflow import cli
-DB_FILES_DIR = ""
+from amworkflow.src.constants.data_model import DeepMapParamModel, MapParamModel
+from amworkflow.src.constants.enums import Label as L
+from amworkflow.src.utils.parser import geom_param_parser, yaml_parser, batch_data_parser
+from amworkflow.src.utils.sanity_check import path_valid_check, dimension_check
+from amworkflow.src.utils.permutator import simple_permutator
+from amworkflow.src.constants.exceptions import NoDataInDatabaseException, InsufficientDataException
+from amworkflow.src.interface.api import amWorkflow as aw
+from amworkflow.src.utils.writer import task_id_creator
+#3rd party lib dependencies
+import numpy as np
+import gmsh
+#built-in dependencies
+import os
+import copy
+import subprocess
+from multiprocessing import Pipe, Process
 
 class BaseWorkflow(object):
     def __init__(self, args):
-        self.raw_args = args
+        #base
+        self.args = args
         self.geometry_spawn: callable
-        print(self.raw_args)
-        print(os.getcwd())
-        self.mpm = MapParamModel
-        self.dmpm = DeepMapParamModel
-        self.parsed_args = cmd_parser(self.raw_args)
-        self.parsed_args_c = self.dmpm(self.parsed_args)
-        self.yaml_dir = self.raw_args.yaml_dir
-        self.import_dir = self.raw_args.import_dir
-        self.yaml_parser = yaml_parser
-        self.namer = namer
-        self.model_name = self.raw_args.name
-        self.model_hashname = self.namer(name_type="hex")
-        self.geom_pointer: int
-        self.shape = []
-        self.mesh_result = []
-        self.name_list = []
-        self.parm_list = []
-        self.hashname_list = []
-        self.mesh_name_list = []
-        self.mesh_hashname_list = []
-        self.db = True
-        self.start_vector = []
-        self.end_vector = []
-        self.num_vector = []
-        self.db_data_collection = {}
-        self.db_del_collection = []
-        self.batch_num = self.parsed_args
+        # data
+        self.geom_data = []
+        self.pm = 0
+        self.mdl = 0
+        self.md5 = ""
+        self.impt_filename = ""
+        self.impt_format = ""
+        self.task_dir = ""
+        self.import_fl = None
+        self.db_data_collect = {"geometry":[],
+                                "occ": [],
+                                "mesh":[]}
+        self.oil = [
+            [0,0,0,1,0],
+            [0,4,1,1,0]
+        ]
+        #signals
+        self.indicator = self.task_handler()
         self.isbatch = False
-        self.linear_deflect = 0.001
-        self.angular_deflect = 0.1
-        self.mesh_t = None
-        self.mesh_n = None
-        self.mesh_s = None
-        self.indicator = task_handler(args=self.raw_args)
-        self.data_init()
-        self.permutation = self.permutator()
-                
-    def data_init(self):
-        match self.indicator[0]: 
-            case 1: #read the step file stored in the database and convert it to an OCC representation.
-                self.linear_deflect = self.raw_args.stl_linear_deflect
-                self.angular_deflect= self.raw_args.stl_angular_deflect
-                self.mesh_t = self.raw_args.mesh_by_thickness
-                self.mesh_n = self.raw_args.mesh_by_layer
-                self.mesh_s = self.raw_args.mesh_size_factor
-                self.isbatch = self.parsed_args[L.BATCH_PARAM.value][L.IS_BATCH.value]
-                impt_format = path_valid_check(self.raw_args.import_dir, format=["stp", "step","stl","STL"])
-                impt_filename = get_filename(self.raw_args.import_dir)
-                match impt_format.lower():
-                    case "stl":
-                        self.import_model = stl_reader(path=self.import_dir)
-                    case "stp":
-                        self.import_model = step_reader(path=self.import_dir)
-                    case "step":
-                        self.import_model = step_reader(path=D.DATABASE_OUTPUT_FILE_PATH.value+impt_filename)
-                match self.indicator[1]:
-                        case 1: # Import file, convert the file to an OCC representation and create a new profile for imported file. 
-                            self.db_data_collection[L.MDL_PROF.value] = {L.MDL_NAME.value: impt_filename} 
-                            # self.batch_data_convert(self.parsed_args[L.GEOM_PARAM.value])
-                        case 2: # remove selected profile and stp file, then quit
-                            #TODO remove the step file and and info in db. 
-                            pass
-            case 2: # yaml file provided
-                self.data = yaml_parser(self.yaml_dir)
-                if L.MDL_NAME.value in self.data["model_profile"]:
-                    self.model_name = self.data["model_profile"][L.MDL_NAME.value]
-                else: raise InsufficientDataException()
-                self.geom_data = self.data[L.GEOM_PARAM.value]
-                self.batch_data_convert(data=self.geom_data)
-                self.param_type = self.geom_data.keys()
-                self.mesh_param = self.data[L.MESH_PARAM.value]
-                self.linear_deflect = self.data[L.STL_PARAM.value][L.LNR_DFT.value]
-                self.angular_deflect = self.data[L.STL_PARAM.value][L.ANG_DFT.value]
-                self.mesh_t = self.data[L.MESH_PARAM.value][L.LYR_TKN.value]
-                self.mesh_n = self.data[L.MESH_PARAM.value][L.LYR_NUM.value]
-                self.mesh_s = self.data[L.MESH_PARAM.value][L.MESH_SIZE_FACTOR.value]
-                if sum(self.num_vector) > 1:
-                    self.isbatch = True
-                    self.batch_data_convert(self.parsed_args[L.GEOM_PARAM.value])
-                else:
-                    self.isbatch = False
-                
-            case 0: # model_name provided
-                self.linear_deflect = self.raw_args.stl_linear_deflect
-                self.angular_deflect= self.raw_args.stl_angular_deflect
-                self.mesh_t = self.raw_args.mesh_by_thickness
-                self.mesh_n = self.raw_args.mesh_by_layer
-                self.mesh_s = self.raw_args.mesh_size_factor
-                self.isbatch = self.parsed_args[L.BATCH_PARAM.value][L.IS_BATCH.value]
-                self.query_list = query_multi_data(table = ModelParameter,
-                                    by_name= self.model_name,
-                                    column_name=L.MDL_NAME.value)
-                self.param_type = list(set(rows[L.PARAM_TP.value] for rows in self.query_list))
-                if self.indicator[1] == 1: # edit mode. edit profile, replace the old and continue with new parameters.
-                    self.db_del_collection.append(ModelParameter,self.param_type)
-                    self.db_delete()
-                    self.param_type = self.parsed_args[L.GEOM_PARAM.value].keys()
-                    value = [{L.PARAM_NAME.value: new_param[i],
-                                  L.MDL_NAME.value: self.impt_filename} for i in range(len(self.param_type))]
-                    self.db_insert(ModelParameter, value)
-                match self.indicator[1]:
-                    case 2:
-                        #remove certain model profile from database
-                        self.db_del_collection.append(ModelProfile, [self.model_name])
-                        self.db_delete()
-                    case 3: # do nothing, fill data into the loaded model.
-                        self.param_type = self.parsed_args[L.GEOM_PARAM.value].keys()
-                    case 4: # Create a new model profile with given parameters.
-                        self.db_data_collection[L.MDL_PROF.value] = [{L.MDL_NAME.value: self.raw_args.name}]
-                        new_param = list(self.parsed_args[L.GEOM_PARAM.value].keys())
-                        print(new_param)
-                        value = [{L.PARAM_NAME.value: new_param[i],
-                                  L.MDL_NAME.value: self.model_name} for i in range(len(new_param))]
-                        self.db_data_collection[L.MDL_PARAM.value] = value
-                        self.param_type = list(self.parsed_args[L.GEOM_PARAM.value].keys())
-                        self.db_insert(ModelProfile, self.db_data_collection[L.MDL_PROF.value])
-                        self.db_insert(ModelParameter, self.db_data_collection[L.MDL_PARAM.value])
-        self.batch_data_convert(data=self.parsed_args[L.GEOM_PARAM.value])
-                
+        self.isimport = False
+        self.onlyimport = False
+        self.cmesh = False
+        self.init_signal = self.data_init()
+        print(self.indicator)
+        #main
+        self.task_id = task_id_creator()
+        
+        self.sequence()
+
+    def sequence(self):
+        match self.init_signal:
+            case 1:
+                print("Done. Quitting...")
+            case 0:
+                self.geom_data = self.geom_param_handler()
+                label = self.args.geom_param
+                value = self.args.geom_param_value
+                self.pm = MapParamModel(label, value)
+                # self.create()
+    
     def create(self) -> None:
-        '''
-        create the real entity of the geometry, then prepare necessary info for sequential processes.
-        Geometries created by this method should be placed in self.shape.
-        '''
-        dimension_check(self.permutation)  
-        is_start_vector = False
-        self.batch_num = batch_num_creator()
-        self.db_data_collection["geometry"] = []
-        for iter_ind, iter_perm in enumerate(self.permutation):
-                if (np.linalg.norm(iter_perm - self.start_vector) == 0):
-                    if is_start_vector == False:
-                        self.geom_process(iter_ind, iter_perm, "dimension-batch")
-                        is_start_vector = True
-                    else:
-                        pass
-                else:
-                    self.geom_process(iter_ind, iter_perm, "dimension-batch")
-        if self.db:
-            self.db_insert(db_model=GeometryFile, data=self.db_data_collection["geometry"])
-            for ind, item in enumerate(self.shape):
-                print(self.angular_deflect)
-                stl_writer(item=item,
-                            item_name=self.hashname_list[ind] + ".stl",
-                            linear_deflection= self.linear_deflect,
-                            angular_deflection= self.angular_deflect,
-                            store_dir=DB_FILES_DIR)
-    
-    def geom_process(self, ind: int, param: list, name_type: str):
-        param = self.mpm(self.param_type, param)
-        tp_geom = self.geometry_spawn(param)
-        self.shape.append(tp_geom)
-        hash_name, filename = self.item_namer(name_type=name_type, ind=ind)
-        append_data = {"batch_num" : self.batch_num,
-                       "geom_hashname" : hash_name,
-                       "filename" : filename,
-                       "model_name": self.model_name}
-        self.db_data_collection["geometry"].append(append_data)
-    
-    def mesh(self):
-        '''
-        mesh the geom created by create()
-        '''   
-        if self.indicator == (2,0):
-            data=[self.mesh_n, self.mesh_t, self.mesh_s]
-            label=list(self.mesh_param.keys())
+        if self.init_signal == 2:
+            self.process_geometry(param=[])
         else:
-            data=[self.mesh_n, self.mesh_t, self.mesh_s]
-            label=[L.LYR_NUM.value,L.LYR_TKN.value,L.MESH_SIZE_FACTOR.value]
-        mesh_param = self.mpm(data=data, label=label)
-        print(mesh_param.dict)
-        print(mesh_param.layer_thickness)
-        is_thickness = True if mesh_param.layer_thickness != None else False
-        size_factor = mesh_param.mesh_size_factor
-        gmsh.initialize()
-        self.db_data_collection["mesh"] = {"xdmf": [],
-                                           "h5": []}
-        for index, item in enumerate(self.shape):
-            
-            if is_thickness:
-                layer_param = mesh_param.layer_thickness
+            if self.isbatch:
+                start_vec = np.array(self.args.geom_param_value)
+                is_start_vector = False
+                for iter_ind, iter_perm in enumerate(self.geom_data):
+                    if (np.linalg.norm(iter_perm - start_vec) == 0):
+                        if is_start_vector == False:
+                            self.process_geometry(param = iter_perm)
+                            is_start_vector = True
+                        else:
+                            pass
+                    else:
+                        self.process_geometry(iter_perm)
             else:
-                layer_param = mesh_param.layer_num
-            model = mesher(item=item,
-                   model_name=self.hashname_list[index],
+                self.process_geometry(param=self.args.geom_param_value)
+        aw.db.insert_data("Task", {"task_id": self.task_id,
+                                   "model_name": self.args.name})
+        aw.db.insert_data("GeometryFile", self.db_data_collect["geometry"], True)
+        if self.args.gen_stp:
+            aw.db.update_data("Task", self.task_id, "task_id", "stp", True)
+            
+    def process_geometry(self, param: list):
+        if self.init_signal == 2:
+            pm = DeepMapParamModel({})
+            pm.imported_model = self.import_fl
+            occ = self.geometry_spawn(pm)
+            hex = aw.tool.namer("hex")
+            name = aw.tool.namer("only_import", param, self.task_id, geom_name=self.args.name)
+            self.isimport = True
+        else:
+            pm = MapParamModel(self.args.geom_param, param)
+            pm.imported_model = self.import_fl
+            occ = self.geometry_spawn(pm)
+            hex = aw.tool.namer("hex")
+            name = aw.tool.namer("dimension-batch", param, self.task_id, self.args.geom_param)
+        data_in = {"filename": name,
+                   "geom_hashname": hex,
+                   "model_name": self.args.name,
+                   "linear_deflection": self.args.stl_linear_deflect,
+                   "angular_deflection": self.args.stl_angular_deflect,
+                   "is_imported": self.isimport,
+                   "task_id": self.task_id}
+        self.task_dir = aw.tool.mk_newdir(self.args.db_file_dir, self.task_id)
+        aw.tool.write_stl(occ, hex, self.args.stl_linear_deflect, self.args.stl_angular_deflect, store_dir=self.task_dir)
+        if self.args.gen_stp:
+            aw.tool.write_step(occ, hex, self.task_dir)
+        self.db_data_collect["geometry"].append(copy.copy(data_in))
+        self.db_data_collect["occ"].append(occ)
+        
+    def mesh(self):
+        is_thickness = True if self.args.mesh_by_thickness is not None else False
+        size_factor = self.args.mesh_size_factor
+        gmsh.initialize()
+        g_info = self.db_data_collect["geometry"]
+        for index, item in enumerate(self.db_data_collect["occ"]):
+            if is_thickness:
+                layer_param = self.args.mesh_by_thickness
+            else:
+                layer_param = self.args.mesh_by_layer
+            model = aw.mesh.mesher(item=item,
+                   model_name=g_info[index]["geom_hashname"],
                    layer_type = is_thickness,
                    layer_param=layer_param,
                    size_factor=size_factor)
-            mesh_hashname = self.namer(name_type="hex")
-            self.mesh_hashname_list.append(mesh_hashname)
-            mesh_name = self.namer(name_type="mesh",
+            mesh_hashname = aw.tool.namer(name_type="hex")
+            mesh_name = aw.tool.namer(name_type="mesh",
                                    is_layer_thickness=True,
                                    layer_param= layer_param,
-                                   geom_name=self.name_list[index][:-4])
-            self.mesh_name_list.append(mesh_name)
-            if self.db:
-                mesh_writer(item = model, 
-                            directory=DB_FILES_DIR, 
-                            filename=self.hashname_list[index],
-                            output_filename = mesh_hashname,
-                            format="xdmf")
-                # mesh_writer(item=model,
-                #             directory=D.USECASE_PATH_PARAMWALL_PATH.value,
-                #             filename=self.hashname_list[index],
-                #             output_filename = mesh_hashname,
-                #             format="msh")
-                xdmf_label = ["xdmf_hashname", "filename", "mesh_size_factor", "layer_thickness", "layer_num", "batch_num","stl_hashname"]
-                h5_label = ["5_hashname", "filename", "batch_num"]
-                db_model_xdmf = self.mpm(xdmf_label)
-                db_model_h5 = self.mpm(h5_label)
-                db_model_xdmf.xdmf_hashname = mesh_hashname
-                db_model_h5.h5_hashname = mesh_hashname
-                db_model_h5.xdmf_hashname = mesh_hashname
-                db_model_xdmf.filename = mesh_name + ".xdmf"
-                db_model_h5.filename = mesh_name + ".h5"
-                db_model_xdmf.mesh_size_factor = mesh_param.mesh_size_factor
-                if is_thickness:
-                    db_model_xdmf.layer_thickness = layer_param
-                else:
-                    db_model_xdmf.layer_num = layer_param
-                db_model_xdmf.batch_num = self.batch_num
-                db_model_xdmf.stl_hashname = self.hashname_list[index]
-                db_model_h5.batch_num = self.batch_num
-                xdmf_collection = self.db_data_collection["mesh"]["xdmf"]
-                h5_collection = self.db_data_collection["mesh"]["h5"]
-                xdmf_collection.append(db_model_xdmf.dict)
-                h5_collection.append(db_model_h5.dict)
+                                   geom_name=g_info[index]["filename"])
+            aw.tool.write_mesh(item = model, 
+                        directory=self.task_dir, 
+                        modelname=g_info[index]["geom_hashname"],
+                        output_filename = mesh_hashname,
+                        format="xdmf")
+            if self.args.gen_msh:
+                aw.tool.write_mesh(item = model, 
+                        directory=self.task_dir, 
+                        modelname=g_info[index]["geom_hashname"],
+                        output_filename = mesh_hashname,
+                        format="msh")
+            if self.args.gen_vtk:
+                aw.tool.write_mesh(item = model, 
+                        directory=self.task_dir, 
+                        modelname=g_info[index]["geom_hashname"],
+                        output_filename = mesh_hashname,
+                        format="vtk")    
+            data_in = {"mesh_hashname":mesh_hashname,
+                       "mesh_size_factor": self.args.mesh_size_factor,
+                       "layer_thickness": self.args.mesh_by_thickness,
+                       "layer_num": self.args.mesh_by_layer,
+                       "task_id": self.task_id,
+                       "geom_hashname": g_info[index]["filename"],
+                       "filename": mesh_name}
+            self.db_data_collect["mesh"].append(copy.copy(data_in))
+        aw.db.insert_data("MeshFile",self.db_data_collect["mesh"], True)
+        aw.db.update_data("Task", self.task_id, "task_id", "xdmf", True)
+        aw.db.update_data("Task", self.task_id, "task_id", "h5", True)
+        if self.args.gen_msh:
+            aw.db.update_data("Task", self.task_id, "task_id", "msh", True)
+        if self.args.gen_vtk:
+            aw.db.update_data("Task", self.task_id, "task_id", "vtk", True)
         gmsh.finalize()
-        if self.db:
-            self.db_insert(XdmfFile, xdmf_collection)
-            self.db_insert(H5File, h5_collection)
         
-    def slice():
-        '''
-        Slice the geometry created by create()
-        '''
-        pass
-    
-    def gcode():
-        '''
-        Create G-code for geometries sliced by slice()
-        '''
-        pass
-    
-
-    
-    def permutator(self):
-        _, perm = simple_permutator(start_point=self.start_vector,
-                                 end_point=self.end_vector,
-                                 num=self.num_vector)
-        return perm
-
-    def db_insert(self, db_model, data) -> None:
-        insert_data(table=db_model, data=data, isbatch=self.isbatch)
-        
-    def db_delete(self) -> None:
-        for request in self.db_del_collection:
-            table = request[0]
-            delete_data(table=table, by_primary_key=request[1], isbatch=self.isbatch)
-    
-    def upload(self):
-        for ind, item in enumerate(self.shape):
-            stl_writer(item=item,
-                        item_name=self.hashname_list[ind] + ".stl",
-                        linear_deflection= self.data.stl_parameter.linear_deflection,
-                        angular_deflection= self.data.stl_parameter.angular_deflection)
-            
-    def download(self):
-        if self.db:
-            if self.batch_num != None:
-                downloader(batch_num=self.batch_num)
+    def data_init(self):
+        if self.args.stl_linear_deflect is None:
+            self.args.stl_linear_deflect = 0.01
+        if self.args.stl_angular_deflect is None:
+            self.args.stl_angular_deflect = 0.1 
+        if ((self.args.mesh_by_layer is not None) or (self.args.mesh_by_thickness is not None)) and (self.args.mesh_size_factor is not None):
+            self.cmesh = True
+        match self.indicator[2]: 
+            case 1: #read the imported file stored in the database and convert it to an OCC representation.
                 
-    def item_namer(self, name_type, ind):
-        hash_name = self.namer(name_type="hex")
-        self.hashname_list.append(hash_name)
-        filename = self.namer(name_type=name_type,
-                    parm_title = self.param_type,
-                    dim_vector=self.permutation[ind],
-                    batch_num=self.batch_num) + ".stl"
-        self.name_list.append(filename)
-        return hash_name, filename
+                # if self.args.geom_param
+                match self.indicator[3]:
+                    case 0: # got the same file in db.
+                        self.load_geom_file()
+            
+                    case 1: # Import file, convert the file to an OCC representation
+                        self.load_geom_file()
+                    case 2: # remove selected profile and geom file, then quit 
+                        aw.db.delete_data("ImportedFile", self.md5)
+                        print(f"File {self.impt_filename}, ID {self.md5} \nin db has been removed successfully...")
+            case 0:
+                match self.indicator[3]:
+                    case 1:
+                        # load stored file into workflows
+                        self.load_geom_file()
+                        
+                    case 2:
+                        # replace geom file from selected profile.
+                        aw.db.update_data("ModelProfile", self.args.name, "model_name", "imported_file_id", self.md5)
+            # case 2: # yaml file provided
+            #     # self.data = DeepMapParamModel(yaml_parser(self.args.yaml_dir))
+            #     data = DeepMapParamModel(yaml_parser(self.args.yaml_dir))
+            #     return data
+        match self.indicator[0]:
+            case 0: # model_name provided
+                match self.indicator[1]:
+                    case 1: #edit selected model
+                        have_data, diff_new, diff_old, query = aw.db.have_data_in_db("ParameterToProfile", "param_name", self.args.geom_param, filter_by=self.args.name, search_column="model_name", search_column2="param_status", filter_by2=1)
+                        new_p = []
+                        active_p = []
+                        if have_data:
+                            print("No new parameters given, quitting...")
+                        else:
+                            for item in diff_old:
+                                aw.db.update_data("ParameterToProfile",by_name = item, on_column="param_name", edit_column="param_status", new_value=False)
+                            for item in diff_new:
+                                q = aw.db.query_data("ModelParameter", by_name=item, column_name="param_name")
+                                if q.empty:
+                                    aw.db.insert_data("ModelParameter", {"param_name": item})
+                                    aw.db.insert_data("ParameterToProfile", {
+                                    "param_name": item,
+                                    "model_name": self.args.name,
+                                    "param_status": True})
+                                    new_p.append(item)
+                                else:
+                                    ex_name = q["param_name"][0]
+                                    aw.db.update_data("ParameterToProfile", ex_name,"param_name","param_status", True )
+                                    active_p.append(item)
+                            print(f"""Deactivate parameter(s):{diff_old} \nActivate parameter(s): {active_p+new_p} \nAdd new parameter(s): {new_p}""")        
+                                    
+                        if self.args.geom_param_value is not None:
+                            self.indicator[4] = 1
+                            
+                    case 2:
+                        #remove certain model profile from database
+                        query_mdl_pfl = aw.db.query_data("ModelProfile", by_name=self.args.name, column_name="model_name")
+                        q_md5 = query_mdl_pfl.imported_file_id[0]
+                        if q_md5 != None:
+                            aw.db.delete_data("ImportedFile", prim_ky=q_md5)
+                        else:
+                            aw.db.delete_data("ModelProfile", prim_ky=self.args.name)
+                        print(f"model {self.args.name} has been removed successfully.")
+                    case 0: # do nothing, fill data into the loaded model.
+                        query = aw.db.query_data("ParameterToProfile", by_name=self.args.name, column_name="model_name", only_for_column="param_name", snd_column_name="param_status", snd_by_name=1)
+                        if (self.args.geom_param is None) and (self.args.geom_param_value is None) and (self.args.import_dir is None):
+                            if self.md5 is None:
+                                print(f"Activated Parameter(s) of model {self.args.name}: {query}")
+                        if self.args.geom_param_value is not None:
+                            self.indicator[4] = 1
+                            self.args.geom_param = query
+                        if self.args.import_dir is not None:
+                            path_valid_check(self.args.import_dir)
+                            
+                    case 3: # Create a new model profile with given parameters.
+                        query_m = aw.db.query_data("ModelProfile", self.args.name, "model_name")
+                        if query_m.empty:
+                            aw.db.insert_data("ModelProfile", {"model_name": self.args.name})
+                        if self.args.geom_param is not None:
+                            input_data = {}
+                            input_data2 = {}
+                            collect = []
+                            collect2 = []
+                            have_data, diff,_,_ = aw.db.have_data_in_db("ModelParameter", "param_name", self.args.geom_param)
+                            if not have_data:
+                                for param in diff:
+                                    input_data.update({"param_name": param})
+                                    collect.append(copy.copy(input_data))
+                                aw.db.insert_data("ModelParameter", collect, True)
+                            for param in self.args.geom_param:
+                                input_data2.update({"param_name": param,
+                                                    "model_name": self.args.name})
+                                collect2.append(copy.copy(input_data2))
+                            aw.db.insert_data("ParameterToProfile", collect2, True)
+                            
+            case 3: # Draft mode.
+                # temp_dir = aw.tool.mk_newdir(self.args.db_dir, "temp")
+                pass
+            
+        if (self.indicator[4] == 1) and (self.args.geom_param_value is not None) and (self.args.geom_param is not None):
+            return 0
+        # elif ((self.indicator[0:2] == [0,4]) or (self.indicator[0:2] == [0,0])) and ((self.indicator[2:4] == [1,0]) or (self.indicator[2:4] == [1,1])):
+        elif self.indicator in self.oil:
+            self.onlyimport = True
+            self.load_geom_file()
+            return 2
+        else:
+            return 1
+            
+    def geom_param_handler(self):
+        if self.args.iter_param is not None:
+            end_vec, num_vec = batch_data_parser(len(self.args.geom_param), self.args.iter_param)
+            start_vec = np.array(self.args.geom_param_value)
+            permutation = simple_permutator(start_vec, end_vec, num_vec, self.args.geom_param)[1]
+            self.isbatch = True
+            return permutation
+        else:
+            return self.args.geom_param_value
     
-    def batch_data_convert(self, data: dict) -> None:
-        for key, item in data.items():
-            self.start_vector.append(item[L.STARTPOINT.value])
-            self.end_vector.append(item[L.ENDPOINT.value])
-            self.num_vector.append(item[L.NUM.value])
-        self.start_vector = np.array(self.start_vector)
-        self.end_vector = np.array(self.end_vector)
-        self.num_vector = np.array(self.num_vector)
-    
-    def create_database_engine(self):
-        pass
-
-def task_handler(args):
-    if args.name != None:
-        result = query_multi_data(ModelProfile, by_name=args.name, column_name=L.MDL_NAME.value, target_column_name=L.MDL_NAME.value)
-        if args.name in result:
-            indicator = (0,0)
-            if args.edit:
-                indicator = (0,1)
-            else:
-                if args.remove:
-                    indicator = (0,2)
+    def task_handler(self):
+        indicator = [0,0,0,0,0]
+        match self.args.mode:
+            case "draft":
+                indicator[0] = 3 #[3,0,0,0,0]
+            case "production":
+                if self.args.yaml_dir is not None:
+                    path_valid_check(self.args.yaml_dir, format=["yml", "yaml"])
+                    self.args = yaml_parser(self.args.yaml_dir)
+                if self.args.name != None:
+                    result = aw.db.query_data("ModelProfile", by_name=self.args.name, column_name=L.MDL_NAME.value)
+                    if not result.empty:
+                        #indicator = [0,0,0,0,0]
+                        self.md5 = result.imported_file_id[0]
+                        query_f = aw.db.query_data("ImportedFile", by_name=self.md5, column_name="md5_id")
+                        if self.md5 is not None:
+                            indicator[2:4] = [0,1] #[0,0,0,1,0]
+                            query_f = aw.db.query_data("ImportedFile", by_name=self.md5, column_name="md5_id")
+                            self.impt_filename = query_f.filename[0]
+                            self.impt_format = path_valid_check(os.path.join(self.args.import_file_dir, self.impt_filename), format=["stp", "step","stl","STL"])
+                            if self.args.replace is not None:
+                                if os.path.isdir(os.path.dirname(self.args.replace)):
+                                    self.args.import_dir = self.args.replace
+                                else:
+                                    if aw.tool.is_md5(self.args.replace):
+                                        indicator[3] = 2 #[0,0,0,2,0]
+                                    else:
+                                        print(f"Got {self.args.replace}")
+                                        raise Exception("Replace: Neither a md5 ID or a file path provided.")
+                        if self.args.edit:
+                            indicator[1] = 1 #[0,1,0,0,0]
+                        elif self.args.remove:
+                            indicator[1] = 2 #[0,2,0,0,0]
+                    elif self.args.geom_param is not None:
+                        indicator[1] = 3 #[0,3,0,0,0]
+                        if self.args.geom_param_value is not None:
+                            indicator[4] = 1 #[0,3,0,0,1]
+                    else:
+                        indicator[1] = 4 #[0,4,0,0,0]
+                    if self.args.import_dir != None:
+                        indicator[2] = 1
+                        self.impt_format = path_valid_check(self.args.import_dir, format=["stp", "step","stl","STL"])
+                        self.isimport = True
+                        self.impt_filename = aw.tool.get_filename(self.args.import_dir)
+                        self.md5 = aw.tool.get_md5(self.args.import_dir)
+                        result = aw.db.query_data("ImportedFile", by_name=self.md5, column_name="md5_id")
+                        if not result.empty:
+                            q_filename = result.filename[0]
+                            self.impt_filename = q_filename
+                            if self.args.remove:
+                                indicator[3] = 2
+                            else:
+                                indicator[3] = 0
+                                print(f"Got same file {q_filename} in db, using existing file now...")
+                        else:
+                            indicator[3] = 1
+                            aw.tool.upload(self.args.import_dir, self.args.import_file_dir)
+                            aw.db.insert_data("ImportedFile",{"filename": self.impt_filename, "md5_id": self.md5})
+                            if (indicator[0:2] != [0,3]) and (indicator[0:2] != [0,4]) and (indicator[0] == 0):
+                                #profile info exists while a file needs to be imported.
+                                aw.db.update_data("ModelProfile",self.args.name, "model_name", "imported_file_id", self.md5)
+                            else:
+                                aw.db.insert_data("ModelProfile", {"model_name": self.args.name,"imported_file_id": self.md5})
+                    # elif (indicator[0:2] == [0,4]) and result.loc["imported_file_id"][0] is not None:
+                    #     self.md5 = result.loc["imported_file_id"][0]
+                    #     query_f = aw.db.query_data("ImportedFile", by_name=self.md5, column_name="md5_id")
+                    #     if not query_f.empty:
+                    #         q_filename = query_f.filename[0]
+                    #         self.impt_filename = q_filename
+                                
+                    # else:
+                    #     raise InsufficientDataException()
+                # elif self.args.yaml_dir is not None:
+                #     indicator = (2,0,0)
                 else:
-                    indicator = (0,3)
+                    q0 = aw.db.query_data("ParameterToProfile")
+                    q1 = aw.db.query_data("ModelParameter")
+                    q2 = aw.db.query_data("ModelProfile")
+                    q3 = aw.db.query_data("Task")
+                    q4 = aw.db.query_data("ImportedFile")
+                    if not q2.empty:
+                        print(f"\033[1mModel Profile\033[0m:\n{q2}\n\033[1mModel Parameters\033[0m:\n{q1}\n\033[1mParameter Properties\033[0m:\n{q0}\n\033[1mImported Files\033[0m:\n{q4}\n\033[1mTask\033[0m:\n{q3}")
+                    else:
+                        print("No model profile found.")
+
+                    # raise InsufficientDataException()
+        return indicator
+    
+    def load_geom_file(self):
+        if self.impt_format in ["stl","STL"]:
+            self.import_fl = aw.tool.read_stl(self.args.import_file_dir + "/" + self.impt_filename)
         else:
-            indicator = (0,4)
-    else:    
-        if args.import_dir != None:
-            impt_format = path_valid_check(args.import_dir, format=["stp", "step","stl","STL"])
-            impt_filename = get_filename(args.import_dir)
-            result = query_multi_data(ModelProfile, by_name=impt_filename, column_name=L.MDL_NAME.value, target_column_name=L.MDL_NAME.value)
-            if impt_filename in result:
-                indicator = (1,0)
-                if args.remove:
-                    indicator = (1,2)
-            else:
-                indicator = (1,1)
-        else:
-            if args.yaml_dir != None:
-                path_valid_check(args.yaml_dir, format=["yml", "yaml"])
-                indicator = (2,0)
-            else:
-                raise InsufficientDataException()
-    return indicator
+            self.import_fl = aw.tool.read_step(self.args.import_file_dir + "/" + self.impt_filename)
+
