+++ amworkflow/src/infrastructure/database/cruds/crud.py
@@ -2,6 +2,7 @@ from amworkflow.src.infrastructure.database.models.model import db_list
 from sqlalchemy import insert
 from sqlalchemy.sql.expression import select
 import pandas as pd
+from amworkflow.src.utils.visualizer import color_background
 
 def insert_data(table: str,
                 data: dict, 
@@ -65,6 +66,7 @@ def query_multi_data(table: str,
             dd.pop("_sa_instance_state", None)
     if target_column_name == None:
         result = pd.DataFrame(result)
+        result.style.applymap(color_background)
     # elif len(result) != 0:
     #     result = result[0]
     return result
@@ -115,3 +117,36 @@ def delete_data(table: str,
 def transcation_rollback():
     from amworkflow.src.infrastructure.database.engine.engine import session
     session.rollback()
+    
+def query_join_tables(table: str, join_column: str, table1: str,  join_column1:str, table2: str = None, join_column2:str = None, filter0: str = None, filter1: str = None, filter2: str = None, on_column_tb: str = None, on_column_tb1: str = None, on_column_tb2: str = None):
+    from amworkflow.src.infrastructure.database.engine.engine import session
+    table = db_list[table]
+    table1 = db_list[table1]
+    on_c0 = getattr(table, join_column)
+    on_c1 = getattr(table1, join_column1)
+    if on_column_tb is not None:
+        column0 = getattr(table, on_column_tb)
+    q = session.query(table).join(table1, on_c0 == on_c1)
+    c = []
+    if filter0 is not None:
+        c0 = filter0 == column0
+        c.append(c0)
+    if filter1 is not None:
+        column1 = getattr(table, on_column_tb1)
+        c1 = filter1 == column1
+        c.append(c1)
+    if table2 is not None:
+        table2= db_list[table2]
+        column2 = getattr(table, on_column_tb2)
+        q = q.join(table2)
+        if filter2 is not None:
+            c2 = filter2 == column2
+            c.append(c2)
+    if len(c) != 0:
+        q.filter(*c)
+    result = [i.__dict__ for i in q.all()]
+    for dd in result:
+        dd.pop("_sa_instance_state", None)
+    
+    return pd.DataFrame(result)
+
